Revert [7907] (chanid from string -> int) change for now - too many hidden breakages due to QStrings operator+ overloading.
git-svn-id: http://svn.mythtv.org/svn/trunk@7925 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythfrontend/manualbox.cpp
@@ -332,7 +332,7 @@ void ManualBox::startClicked(void)
     progInfo.subtitle = m_subtitle->text();
     progInfo.description = m_descString;
     progInfo.category = m_categoryString;
-    progInfo.chanid = m_chanidString.toUInt();
+    progInfo.chanid = m_chanidString;
 
     QDateTime now = QDateTime::currentDateTime();
     // Make sure we set the seconds to 0
